//check if civicontribute is enable before deleting..
git-svn-id: http://svn.civicrm.org/civicrm/branches/v1.7@9014 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ CRM/Contact/BAO/Contact.php
@@ -1314,7 +1314,10 @@ WHERE civicrm_contact.id IN $idString ";
         
         CRM_Contact_BAO_Relationship::deleteContact( $id );
 
-        CRM_Contribute_BAO_Contribution::deleteContact( $id );
+        if ( CRM_Core_Permission::access( 'CiviContribute' ) ) {
+            require_once 'CRM/Contribute/BAO/Contribution.php';
+            CRM_Contribute_BAO_Contribution::deleteContact( $id );
+        }
         
         if ( CRM_Core_Permission::access( 'CiviMember' ) ) {
             require_once 'CRM/Member/BAO/Membership.php';
